requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
blessed==1.18.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
chronological==0.1.1
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
et-xmlfile==1.1.0
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==21.8.0
loguru==0.5.3
lxml==4.6.3
MarkupSafe==2.0.1
measurement==3.2.0
mpmath==1.2.1
multidict==5.1.0
myfitnesspal==1.16.4
numpy==1.21.1
openai==0.10.2
openpyxl==3.0.7
pandas==1.3.1
pandas-stubs==1.2.0.1
Pygments==2.9.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rich==9.13.0
six==1.16.0
sympy==1.8
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
yarl==1.6.3
